Inclusão validações de preenchimento e regras de negocio

+++ routes/web.php
@@ -23,34 +23,35 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::group(['middleware' => 'auth'], function () {
+
     Route::get('/', [HomeController::class, 'index'])->name('index');
     Route::post('/logout', [AuthenticationController::class, 'logout'])->name('logout');
 });
 
 Route::prefix('movies')->group(function () {
-    Route::get('/', [MovieController::class, 'index'])->name('movies-index');
-    Route::get('/{id}/edit', [MovieController::class, 'edit'])->where('id', '[0-9]+')->name('movies-edit');
-    Route::get('/create', [MovieController::class, 'create'])->name('movies-create');
-    Route::post('/', [MovieController::class, 'store'])->name('movies-store');
-    Route::put('/{id}', [MovieController::class, 'update'])->name('movies-update');
-    Route::delete('/{id}', [MovieController::class, 'destroy'])->where('id', '[0-9]+')->name('movies-destroy');
+    Route::get('/', [MovieController::class, 'index'])->name('movies-index')->middleware('auth');
+    Route::get('/{id}/edit', [MovieController::class, 'edit'])->where('id', '[0-9]+')->name('movies-edit')->middleware('auth');
+    Route::get('/create', [MovieController::class, 'create'])->name('movies-create')->middleware('auth');
+    Route::delete('/{id}', [MovieController::class, 'destroy'])->where('id', '[0-9]+')->name('movies-destroy')->middleware('auth');
+    Route::put('/{id}', [MovieController::class, 'update'])->name('movies-update')->middleware('auth')->middleware('auth');
+    Route::post('/', [MovieController::class, 'store'])->name('movies-store')->middleware('auth')->middleware('auth');
 });
 
 Route::prefix('rooms')->group(function () {
-    Route::get('/', [RoomController::class, 'index'])->name('rooms-index');
-    Route::get('/create', [RoomController::class, 'create'])->name('rooms-create');
-    Route::post('/', [RoomController::class, 'store'])->name('rooms-store');
-    Route::put('/{id}', [RoomController::class, 'update'])->name('rooms-update');
-    Route::get('/{id}/edit', [RoomController::class, 'edit'])->where('id', '[0-9]+')->name('rooms-edit');
-    Route::delete('/{id}', [RoomController::class, 'destroy'])->where('id', '[0-9]+')->name('rooms-destroy');
+    Route::get('/', [RoomController::class, 'index'])->name('rooms-index')->middleware('auth');
+    Route::get('/create', [RoomController::class, 'create'])->name('rooms-create')->middleware('auth');
+    Route::get('/{id}/edit', [RoomController::class, 'edit'])->where('id', '[0-9]+')->name('rooms-edit')->middleware('auth');
+    Route::delete('/{id}', [RoomController::class, 'destroy'])->where('id', '[0-9]+')->name('rooms-destroy')->middleware('auth');
+    Route::put('/{id}', [RoomController::class, 'update'])->name('rooms-update')->middleware('auth')->middleware('auth');
+    Route::post('/', [RoomController::class, 'store'])->name('rooms-store')->middleware('auth')->middleware('auth');
 });
 
 Route::prefix('/sessions')->group(function () {
-    Route::get('/create', [SessionController::class, 'create'])->name('sessions-create');
-    Route::post('/', [SessionController::class, 'store'])->name('sessions-store');
-    Route::get('/edit/{id}', [SessionController::class, 'edit'])->where('id', '[0-9]+')->name('sessions-edit');
-    Route::put('/{id}', [SessionController::class, 'update'])->name('sessions-update');
-    Route::delete('/{id}', [SessionController::class, 'destroy'])->where('id', '[0-9]+')->name('sessions-destroy');
+    Route::get('/create', [SessionController::class, 'create'])->name('sessions-create')->middleware('auth');
+    Route::get('/edit/{id}', [SessionController::class, 'edit'])->where('id', '[0-9]+')->name('sessions-edit')->middleware('auth');
+    Route::delete('/{id}', [SessionController::class, 'destroy'])->where('id', '[0-9]+')->name('sessions-destroy')->middleware('auth');
+    Route::post('/', [SessionController::class, 'store'])->name('sessions-store')->middleware('auth');
+    Route::put('/{id}', [SessionController::class, 'update'])->name('sessions-update')->middleware('auth')->middleware('auth');
 });
